refactor: move common functions to eyes_selenium_utils
